Move rpc tests from rpc_test packages to rpc so internal state can be examined
