+++ core/node/rpc/new_node_test.go
@@ -1,4 +1,4 @@
-package rpc_test
+package rpc
 
 import (
 	"context"
