+++ core/node/rpc/standby_test.go
@@ -1,4 +1,4 @@
-package rpc_test
+package rpc
 
 import (
 	"encoding/json"
